refactored code and customize index file
