+++ components/Author/Authors.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { faker } from "@faker-js/faker";
-import Story from "./Story";
+import Story from "../Profile/Story";
 
 const Authors = () => {
   const [stories, setStories] = useState([]);
