+++ pages/index.tsx
@@ -1,84 +1,56 @@
 import type { NextPage } from "next";
 import Head from "next/head";
 import Image from "next/image";
+import Link from "next/link";
 
 const Home: NextPage = () => {
   return (
-    <div className="flex min-h-screen flex-col items-center justify-center py-2">
+    <div className="flex min-h-screen flex-col text-center items-center justify-center py-2">
       <Head>
-        <title>Create Next App</title>
+        <title>Wrider</title>
         <link rel="icon" href="/favicon.ico" />
       </Head>
-
-      <main className="flex w-full flex-1 flex-col items-center justify-center px-20 text-center">
-        <h1 className="text-6xl font-bold">
-          Welcome to{" "}
-          <a className="text-blue-600" href="https://nextjs.org">
-            Next.js!
-          </a>
-        </h1>
-
-        <p className="mt-3 text-2xl">
-          Get started by editing{" "}
-          <code className="rounded-md bg-gray-100 p-3 font-mono text-lg">
-            pages/index.tsx
-          </code>
+      <main>
+        <h1 className="text-8xl my-2 font-bold ">Wrider</h1>
+        <p className="my-2 font-semibold text-xl  text-blue-500">
+          This web application is still in production.🕐🕐
         </p>
-
-        <div className="mt-6 flex max-w-4xl flex-wrap items-center justify-around sm:w-full">
-          <a
-            href="https://nextjs.org/docs"
-            className="mt-6 w-96 rounded-xl border p-6 text-left hover:text-blue-600 focus:text-blue-600"
-          >
-            <h3 className="text-2xl font-bold">Documentation &rarr;</h3>
-            <p className="mt-4 text-xl">
-              Find in-depth information about Next.js features and its API.
-            </p>
-          </a>
-
-          <a
-            href="https://nextjs.org/learn"
-            className="mt-6 w-96 rounded-xl border p-6 text-left hover:text-blue-600 focus:text-blue-600"
-          >
-            <h3 className="text-2xl font-bold">Learn &rarr;</h3>
-            <p className="mt-4 text-xl">
-              Learn about Next.js in an interactive course with quizzes!
-            </p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/canary/examples"
-            className="mt-6 w-96 rounded-xl border p-6 text-left hover:text-blue-600 focus:text-blue-600"
-          >
-            <h3 className="text-2xl font-bold">Examples &rarr;</h3>
-            <p className="mt-4 text-xl">
-              Discover and deploy boilerplate example Next.js projects.
-            </p>
-          </a>
-
-          <a
-            href="https://vercel.com/import?filter=next.js&utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className="mt-6 w-96 rounded-xl border p-6 text-left hover:text-blue-600 focus:text-blue-600"
-          >
-            <h3 className="text-2xl font-bold">Deploy &rarr;</h3>
-            <p className="mt-4 text-xl">
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
+        <span className="text-xl ">
+          Some of the page you can visit in meantime to see the production. 🚀🚀
+        </span>
+        <div className="flex space-x-2  my-6">
+          <Link href="/login">
+            <span className="cursor-pointer font-semibold border-r-2 px-6 border-black hover:text-blue-500">
+              Login
+            </span>
+          </Link>
+          <Link href="/register">
+            <span className="cursor-pointer font-semibold border-r-2 px-6 border-black hover:text-blue-500">
+              Register
+            </span>
+          </Link>
+          <Link href="/dashboard">
+            <span className="cursor-pointer font-semibold border-r-2 px-6 border-black hover:text-blue-500">
+              Dashboard
+            </span>
+          </Link>
+          <Link href="/profile">
+            <span className="cursor-pointer font-semibold border-r-2 px-6 border-black hover:text-blue-500">
+              Profile
+            </span>
+          </Link>
+          <Link href="/blog">
+            <span className="cursor-pointer font-semibold border-r-2 px-6 border-black hover:text-blue-500">
+              Blog
+            </span>
+          </Link>
+          <Link href="/blog">
+            <span className="cursor-pointer font-semibold  border-black hover:text-blue-500">
+              Blog writting page.
+            </span>
+          </Link>
         </div>
       </main>
-
-      <footer className="flex h-24 w-full items-center justify-center border-t">
-        <a
-          className="flex items-center justify-center gap-2"
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{" "}
-          <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-        </a>
-      </footer>
     </div>
   );
 };
