+++ tsconfig.json
@@ -19,9 +19,9 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
-    "components/YourArticle.jsx",
-    "components/YourArticle.jsx",
-    "components/BlogSection.jsx",
+    "components/Blog/YourArticle.jsx",
+    "components/Blog/YourArticle.jsx",
+    "components/Blog/BlogSection.jsx",
     "pages/writtingpage.tsx"
   ],
   "exclude": ["node_modules"]
